aider: Remove unnecessary line that updates `qrCode` in `handleMessage` method.
# Aider chat conversation:

USER: I don't think the first line of handling the qr_update message is necessary? This one: this.qrCode = message.qr_code_url + '?' + new Date().getTime();

ASSISTANT: You're correct. The line `this.qrCode = message.qr_code_url + '?' + new Date().getTime();` in the `handleMessage` method of `App.vue` is not necessary because the QR code URL is updated in the next line with the `refreshImage` method of the `QRCode` component.

Here's the change:

```javascript
frontend/src/App.vue
<<<<<<< ORIGINAL
    handleMessage(event) {
      const message = JSON.parse(event.data);
      if (message.type === 'QR_UPDATE') {
        this.qrCode = message.qr_code_url + '?' + new Date().getTime();
        this.$refs.qrCodeComponent.refreshImage();
      } else if (message.type === 'FILES_RETRIEVED') {
        this.results = message.data;
        this.view = 'results';
      } else if (message.type === 'PROGRESS_UPDATE') {
        this.progress = message.progress;
        this.view = 'progress';
      } else if (message.type === 'ERROR') {
        this.error = message.error;
        this.view = 'button';
      }
    },
=======
    handleMessage(event) {
      const message = JSON.parse(event.data);
      if (message.type === 'QR_UPDATE') {
        this.$refs.qrCodeComponent.refreshImage();
      } else if (message.type === 'FILES_RETRIEVED') {
        this.results = message.data;
        this.view = 'results';
      } else if (message.type === 'PROGRESS_UPDATE') {
        this.progress = message.progress;
        this.view = 'progress';
      } else if (message.type === 'ERROR') {
        this.error = message.error;
        this.view = 'button';
      }
    },
>>>>>>> UPDATED
```

This change removes the unnecessary line that updates the `qrCode` data property in the `handleMessage` method. The `refreshImage` method of the `QRCode` component will handle the QR code URL update.