aider: Replaced `puts` statements with a yield statement that returns a hash with `:type` as `LOG` and the message as `:data` in `bank_buster.rb`.
# Aider chat conversation:

USER: We need to make sure that all code in BankBuster is adapted for the frontend use case, instead of the old use case of running it as a script in iTerm. Make changes one at a time. I think it makes sense to yield with :type as LOG and the message as :data maybe.

ASSISTANT: To adapt the BankBuster code for frontend use, we need to modify the way it logs information. Instead of printing to the console, we should yield a hash with `:type` as `LOG` and the message as `:data`. This will allow the frontend to handle and display these log messages appropriately.

The files we need to modify are:

1. `bank_buster.rb`
2. `handlers/bank_buster_handler.rb`

Let's start with the `bank_buster.rb` file.

The first change is to replace all `puts` statements with a yield statement that returns a hash with `:type` as `LOG` and the message as `:data`.

Here is the first change:

```ruby
bank_buster.rb
<<<<<<< HEAD
    puts 'Trying to identify cookie accept button...'.gray
=======
    yield({ type: 'LOG', data: 'Trying to identify cookie accept button...' }) if block_given?
>>>>>>> updated
```

Please confirm this change so we can proceed with the rest.